Early Warning for Disease Clusters

early_warning_cluster() can detect so-called disease clusters. Use has_clusters() to return TRUE or FALSE on its output, or use format() to format the output.

early_warning_cluster(
  df,
  ...,
  group_by = NULL,
  column_date = NULL,
  column_patientid = NULL,
  based_on_historic_maximum = FALSE,
  minimum_ongoing_days = 2,
  threshold_percentile = 97.5,
  remove_outliers = TRUE,
  remove_outliers_coefficient = 1.5,
  moving_average_days = 5,
  moving_average_side = "centre",
  case_free_days = 14
)

has_clusters(x, minimum = 1)

Arguments

df

data set, this must be a data set containing only positive results. The most bare-bone data set to use must have at least a date column and patient column. Do not summarise on patient IDs, this will happen automatically.

...

filter arguments to define the data to study, e.g., jaar == 2023. All other observations will be used as a control.

group_by

column to use in group_by(), e.g., a microorganism name

column_date

name of the column to use for dates, will take the first date column if left blank

column_patientid

name of the column to use for patient IDs, will take the first column names resembling "patient|patid" if left blank

based_on_historic_maximum

logical to indicate whether the percentile must be based on the maximum of previous years. With FALSE (default), the percentile will be based on all historic data points.

minimum_ongoing_days

number of days that must have passed above the set percentile before detecting the elevation as cluster

threshold_percentile

threshold to set

remove_outliers

logical to indicate whether outliers must be removed before threshold determination

remove_outliers_coefficient

coefficient to use for outlier determination

moving_average_days

number of days to set in moving_average()

moving_average_side

side of days to set in moving_average()

case_free_days

number of days to set in get_episode()

x

output of early_warning_cluster()

minimum

minimum number of clusters to check against, defaults to 1

Details

A (disease) cluster is defined as an unusually large aggregation of disease events in time or space (ATSDR, 2008). They are common, particularly in large populations. From a statistical standpoint, it is nearly inevitable that some clusters of chronic diseases will emerge within various communities, be it schools, church groups, social circles, or neighborhoods. Initially, these clusters are often perceived as products of specific, predictable processes rather than random occurrences in a particular location, akin to a coin toss.

Whether a (suspected) cluster corresponds to an actual increase of disease in the area, needs to be assessed by an epidemiologists or biostatistician (ATSDR, 2008).

Examples

cases <- data.frame(date = sample(seq(as.Date("2018-01-01"),
                                      as.Date("2022-12-31"),
                                      "1 day"),
                                  size = 300),
                    patient = sample(LETTERS, size = 300, replace = TRUE))
check <- early_warning_cluster(cases,
                               date >= "2022-01-01",
                               threshold_percentile = 0.99)
#> Using column 'date' for dates
#> Using column 'patient' for patient IDs
#> Warning: There were 30 warnings in `mutate()`.
#> The first warning was:
#> ℹ In argument: `max_ma_5c = max(...)`.
#> ℹ In group 6: `month_day = 1972-01-10`.
#> Caused by warning in `base::max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 29 remaining warnings.

has_clusters(check)
#> [1] FALSE
check
#> => Detected no disease clusters based on an outlier-free history (coeff = 1.5) and the 99th pct lasting for >=2 days.

check2 <- early_warning_cluster(cases,
                                date >= "2022-01-01",
                                threshold_percentile = 0.75)
#> Using column 'date' for dates
#> Using column 'patient' for patient IDs
#> Warning: There were 30 warnings in `mutate()`.
#> The first warning was:
#> ℹ In argument: `max_ma_5c = max(...)`.
#> ℹ In group 6: `month_day = 1972-01-10`.
#> Caused by warning in `base::max()`:
#> ! no non-missing arguments to max; returning -Inf
#> ℹ Run `dplyr::last_dplyr_warnings()` to see the 29 remaining warnings.

check2
#> => Detected no disease clusters based on an outlier-free history (coeff = 1.5) and the 75th pct lasting for >=2 days.
check |> format()
#> [1] cluster      first_day    last_day     ongoing_days total_cases 
#> <0 rows> (or 0-length row.names)

check2 |> has_clusters()
#> [1] FALSE
check2 |> has_clusters(15)
#> [1] FALSE

check |> unclass()
#> $clusters
#> [1] date         cases        cluster      ongoing_days
#> <0 rows> (or 0-length row.names)
#> 
#> $details
#> # A tibble: 300 × 7
#>     year month_day  in_scope cases ma_5c max_ma_5c ma_5c_pct_outscope
#>    <int> <date>     <lgl>    <int> <dbl>     <dbl>              <dbl>
#>  1  2018 1972-01-02 FALSE        1     1         1                  1
#>  2  2018 1972-01-15 FALSE        1     1         1                  1
#>  3  2018 1972-01-17 FALSE        1     1         1                  1
#>  4  2018 1972-01-19 FALSE        1     1         1                  1
#>  5  2018 1972-02-03 FALSE        1     1         1                  1
#>  6  2018 1972-02-06 FALSE        1     1         1                  1
#>  7  2018 1972-02-07 FALSE        1     1         1                  1
#>  8  2018 1972-02-08 FALSE        1     1         1                  1
#>  9  2018 1972-02-09 FALSE        1     1         1                  1
#> 10  2018 1972-02-12 FALSE        1     1         1                  1
#> # ℹ 290 more rows
#> 
#> attr(,"threshold_percentile")
#> [1] 99
#> attr(,"based_on_historic_maximum")
#> [1] FALSE
#> attr(,"remove_outliers")
#> [1] TRUE
#> attr(,"remove_outliers_coefficient")
#> [1] 1.5
#> attr(,"moving_average_days")
#> [1] 5
#> attr(,"minimum_ongoing_days")
#> [1] 2