Early Warning for Outbreaks

Feed in positive cases, and this function will determine potential outbreaks. Use format() to format the output.

early_warning_outbreak(
  df,
  ...,
  group_by = NULL,
  column_date = NULL,
  column_patientid = NULL,
  minimum_ongoing_days = 2,
  threshold_percentile = 97.5,
  remove_outliers = TRUE,
  remove_outliers_coefficient = 1.5,
  moving_average_days = 5,
  moving_average_side = "centre",
  case_free_days = 14
)

Arguments

df

data set, this must be a data set containing only positive results. The most bare-bone data set to use must have at least a date column and patient column. Do not summarise on patient IDs, this will happen automatically.

...

filter arguments to set apart the data to compare, e.g., jaar == 2023

group_by

column to use in group_by(), e.g., testcode or BepalingCode

column_date

name of the column to use for dates, will take the first date column if left blank

column_patientid

name of the column to use for patient IDs, will take the first column names resembling "patient|patid" if left blank

threshold_percentile

threshold to set

remove_outliers

logical to indicate whether outliers must be removed before threshold determination

remove_outliers_coefficient

coefficient to use for outlier determination

moving_average_days

number of days to set in moving_average()

moving_average_side

side of days to set in moving_average()

case_free_days

number of days to set in get_episode()

Examples

cases <- data.frame(date = sample(seq(as.Date("2018-01-01"),
                                      as.Date("2022-12-31"),
                                      "1 day"),
                                  size = 300),
                    patient = sample(LETTERS, size = 300, replace = TRUE))
check <- early_warning_outbreak(cases,
                                date >= "2022-01-01")
#> Using column 'date' for dates
#> Using column 'patient' for patient IDs

format(check)
#> Error in summarise(group_by(x$outbreaks, episode), first_day = min(date,     na.rm = TRUE), last_day = max(date, na.rm = TRUE), ongoing_days = max(ongoing_days,     na.rm = TRUE), total_cases = sum(cases, na.rm = TRUE)): ℹ In argument: `ongoing_days = max(ongoing_days, na.rm = TRUE)`.
#> ℹ In group 1: `episode = 3`.
#> Caused by error:
#> ! object 'ongoing_days' not found

check2 <- early_warning_outbreak(cases,
                                date >= "2022-01-01",
                                threshold_percentile = 0.75)
#> Using column 'date' for dates
#> Using column 'patient' for patient IDs

print(check2)
#> Error in summarise(group_by(x$outbreaks, episode), first_day = min(date,     na.rm = TRUE), last_day = max(date, na.rm = TRUE), ongoing_days = max(ongoing_days,     na.rm = TRUE), total_cases = sum(cases, na.rm = TRUE)): ℹ In argument: `ongoing_days = max(ongoing_days, na.rm = TRUE)`.
#> ℹ In group 1: `episode = 1`.
#> Caused by error:
#> ! object 'ongoing_days' not found

unclass(check2)
#> $outbreaks
#> # A tibble: 65 × 11
#> # Groups:   episode [7]
#>    date       in_scope cases ma_5c  year month_day  is_outlier
#>    <date>     <lgl>    <int> <dbl> <int> <date>     <lgl>     
#>  1 2022-01-11 TRUE         1   0.2  2022 1972-01-11 FALSE     
#>  2 2022-01-22 TRUE         1   0.4  2022 1972-01-22 FALSE     
#>  3 2022-01-23 TRUE         1   0.6  2022 1972-01-23 TRUE      
#>  4 2022-01-25 TRUE         1   0.4  2022 1972-01-25 FALSE     
#>  5 2022-01-29 TRUE         1   0.2  2022 1972-01-29 FALSE     
#>  6 2022-02-17 TRUE         1   0.2  2022 1972-02-17 FALSE     
#>  7 2022-02-22 TRUE         1   0.4  2022 1972-02-22 FALSE     
#>  8 2022-02-24 TRUE         1   0.4  2022 1972-02-24 FALSE     
#>  9 2022-02-27 TRUE         1   0.2  2022 1972-02-27 FALSE     
#> 10 2022-03-03 TRUE         1   0.2  2022 1972-03-03 FALSE     
#> # ℹ 55 more rows
#> # ℹ 4 more variables: ma_5c_pct_outscope <dbl>, week <chr>, cases_week <int>,
#> #   episode <int>
#> 
#> $details
#> # A tibble: 1,826 × 6
#>     year month_day  in_scope cases ma_5c ma_5c_pct_outscope
#>    <int> <date>     <lgl>    <int> <dbl>              <dbl>
#>  1  2018 1972-01-01 FALSE        1 0.333                  0
#>  2  2018 1972-01-02 FALSE        0 0.25                   0
#>  3  2018 1972-01-03 FALSE        0 0.2                    0
#>  4  2018 1972-01-04 FALSE        0 0                      0
#>  5  2018 1972-01-05 FALSE        0 0                      0
#>  6  2018 1972-01-06 FALSE        0 0                      0
#>  7  2018 1972-01-07 FALSE        0 0                      0
#>  8  2018 1972-01-08 FALSE        0 0                      0
#>  9  2018 1972-01-09 FALSE        0 0                      0
#> 10  2018 1972-01-10 FALSE        0 0                      0
#> # ℹ 1,816 more rows
#> 
#> attr(,"threshold_percentile")
#> [1] 0.75
#> attr(,"remove_outliers_coefficient")
#> [1] 1.5
#> attr(,"moving_average_days")
#> [1] 5
#> attr(,"minimum_ongoing_days")
#> [1] 2