Early Warning for Disease Clusters

early_warning_cluster() can detect so-called disease clusters. Use has_clusters() to return TRUE or FALSE on its output, or use format() to format the output.

early_warning_cluster(
  df,
  ...,
  group_by = NULL,
  column_date = NULL,
  column_patientid = NULL,
  based_on_historic_maximum = FALSE,
  minimum_ongoing_days = 2,
  threshold_percentile = 97.5,
  remove_outliers = TRUE,
  remove_outliers_coefficient = 1.5,
  moving_average_days = 5,
  moving_average_side = "centre",
  case_free_days = 14
)

n_clusters(x)

has_clusters(x, n = 1)

has_ongoing_cluster(x, reference = Sys.Date() - 1)

Arguments

df

data set, this must be a data set containing only positive results. The most bare-bone data set to use must have at least a date column and patient column. Do not summarise on patient IDs, this will happen automatically.

...

filter arguments to define the data to study, e.g., jaar == 2023. All other observations will be used as a control.

group_by

column to use in group_by(), e.g., a microorganism name

column_date

name of the column to use for dates, will take the first date column if left blank

column_patientid

name of the column to use for patient IDs, will take the first column names resembling "patient|patid" if left blank

based_on_historic_maximum

logical to indicate whether the percentile must be based on the maximum of previous years. With FALSE (default), the percentile will be based on all historic data points.

minimum_ongoing_days

number of days that must have passed above the set percentile before detecting the elevation as cluster

threshold_percentile

threshold to set

remove_outliers

logical to indicate whether outliers must be removed before threshold determination

remove_outliers_coefficient

coefficient to use for outlier determination

moving_average_days

number of days to set in moving_average()

moving_average_side

side of days to set in moving_average()

case_free_days

number of days to set in get_episode()

x

output of early_warning_cluster()

n

number of clusters, defaults to 1

reference

date to test whether any of the clusters currently has this date in it

Details

A (disease) cluster is defined as an unusually large aggregation of disease events in time or space (ATSDR, 2008). They are common, particularly in large populations. From a statistical standpoint, it is nearly inevitable that some clusters of chronic diseases will emerge within various communities, be it schools, church groups, social circles, or neighborhoods. Initially, these clusters are often perceived as products of specific, predictable processes rather than random occurrences in a particular location, akin to a coin toss.

Whether a (suspected) cluster corresponds to an actual increase of disease in the area, needs to be assessed by an epidemiologists or biostatistician (ATSDR, 2008).

Examples

cases <- data.frame(date = sample(seq(as.Date("2015-01-01"),
                                      as.Date("2022-12-31"),
                                      "1 day"),
                                  size = 300),
                    patient = sample(LETTERS, size = 300, replace = TRUE))

------------------------------------------------------------

check <- early_warning_cluster(cases,
                               date >= "2022-01-01",
                               threshold_percentile = 0.99)
#> Using column 'date' for dates
#> Using column 'patient' for patient IDs
#> Error: object 'check' not found

has_clusters(check)
#> Error in eval(expr, envir, enclos): object 'check' not found
check
#> Error in eval(expr, envir, enclos): object 'check' not found

check2 <- early_warning_cluster(cases,
                                date >= "2022-01-01",
                                threshold_percentile = 0.75)
#> Using column 'date' for dates
#> Using column 'patient' for patient IDs

check2
#> => Detected 5 disease clusters with a total of 14 cases based on an outlier-free history (coeff = 1.5) and the 75th pct lasting for >=2 days.
#>    - The 1st cluster has 4 cases between 2 and 8 februari 2022
#>    - The 2nd cluster has 2 cases between 15 and 17 maart 2022
#>    - The 3rd cluster has 2 cases between 20 and 22 april 2022
#>    - The 4th cluster has 4 cases between 7 and 23 mei 2022
#>    - The 5th cluster has 2 cases between 23 and 25 december 2022
#> Use plot2() to plot the results.
check2 |> format()
#> # A tibble: 5 × 5
#>   cluster first_day  last_day   ongoing_days total_cases
#>     <int> <date>     <date>            <int>       <int>
#> 1       1 2022-02-02 2022-02-08            4           4
#> 2       2 2022-03-15 2022-03-17            2           2
#> 3       3 2022-04-20 2022-04-22            2           2
#> 4       4 2022-05-07 2022-05-23            4           4
#> 5       5 2022-12-23 2022-12-25            2           2

check2 |> n_clusters()
#> [1] 5
check2 |> has_clusters()
#> [1] TRUE
check2 |> has_clusters(15)
#> [1] FALSE

check2 |> has_ongoing_cluster("2022-06-01")
#> [1] FALSE

check |> unclass()
#> Error in eval(expr, envir, enclos): object 'check' not found